fix(zones): JSON-safe relaxed ext_atr + persist fallback; add breakout-quality filter
Two fixes:
1. ext_atr in the relaxed-eligible state was a numpy float (from ATR/EMA math),
   which is not JSON-serializable — so the engine_kv upsert threw on every scan
   that had an eligible ticker, FREEZING the entire zone snapshot (armed display
   + relaxed badge stuck since the pre-#73 write). Cast to plain float, and make
   _persist_zones retry core-only if the optional `relaxed` payload ever fails,
   so a bad value can never again block the zone snapshot.
2. Breakout-quality filter: the confirming 1m bar must close in the top 60% of
   its range (bottom 40% for SHORT). Rejects red rejection bars that poke the
   level then close weak — the MARA bull-trap pattern. Applied to compression,
   pullback, and swing via bar high/low passed from on_bar.

+++ engine/entry_gate.py
@@ -252,7 +252,7 @@ def momentum_relaxed_state(df_entry: pd.DataFrame, price: float) -> dict:
     if atr <= 0:
         return out
     dev = (price - ema21) / atr
-    out["ext_atr"] = round(dev, 2)
+    out["ext_atr"] = float(round(dev, 2))   # plain float — numpy types break JSON upsert
     strong_vol = False
     if "volume" in df_entry.columns and len(df_entry) >= 11:
         rv = float(df_entry["volume"].iloc[-1])
